tests/server/pyproject.toml
[tool.isort]
profile = 'black'
multi_line_output = 3

[tool.mypy]
plugins = "numpy.typing.mypy_plugin"
show_error_codes = true
strict = true

[[tool.mypy.overrides]]
module = [
]
ignore_missing_imports = true

[tool.poetry]
name = "src"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "AGPL"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pytest = "^7.2.0"
httpx = "^0.23.3"
pytest-asyncio = "^0.20.3"
grpcio = "^1.51.3"
protobuf = "^4.22.0"

[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
isort = "^5.10.1"
black = "^22.10.0"
python-lsp-server = "^1.7.0"
mypy-ls = "^0.5.1"
pyls-isort = "^0.2.2"
python-lsp-black = "^1.2.1"
pyflakes = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
